src: Add core C# scripts for the experiment
This commit adds the complete set of C# scripts that implement the "Strangest Clock Experiment".

It includes the logic for both the Traditional and Lazy Update models, as well as all manager components (ExperimentController, ObjectManager, DataCollector, etc.) required to run the simulation and collect data as detailed in the setup manual.