docs: Add "Virtual Faraday Cage" hypothesis paper
This paper proposes the "Virtual Faraday Cage" hypothesis as a technologically-grounded and non-disastrous solution to the Fermi Paradox.

It builds upon the proven feasibility of the Lazy Update World Model, arguing that advanced civilizations would rationally choose to migrate into hyper-efficient virtual worlds, thus becoming silent to the physical universe. This redefines the "Great Filter" as a consequence of ultimate success, not failure.